refactor: Remove ChatOllamaTools in favour of ChatOllama
