+++ packages/langchain/README.md
@@ -110,7 +110,6 @@ The following integrations are available in LangChain.dart:
 | [ChatGoogleGenerativeAI](https://langchaindart.dev/#/modules/model_io/models/chat_models/integrations/googleai)         | [langchain_google](https://pub.dev/packages/langchain_google)       | ✔         | ✔           | ✔         | [Google AI for Developers API](https://ai.google.dev) (aka Gemini API)                                                                                                                                                                                                                                                                                                                                                                       |
 | [ChatMistralAI](https://langchaindart.dev/#/modules/model_io/models/chat_models/integrations/mistralai)                 | [langchain_mistralai](https://pub.dev/packages/langchain_mistralai) | ✔         |             |           | [Mistral Chat API](https://ollama.ai)                                                                                                                                                                                                                                                                                                                                                                                                        |
 | [ChatOllama](https://langchaindart.dev/#/modules/model_io/models/chat_models/integrations/ollama)                       | [langchain_ollama](https://pub.dev/packages/langchain_ollama)       | ✔         | ✔           | ✔         | [Ollama Chat API](https://ollama.ai)                                                                                                                                                                                                                                                                                                                                                                                                         |
-| [ChatOllamaTools](https://langchaindart.dev/#/modules/model_io/models/chat_models/integrations/ollama_tools)            | [langchain_ollama](https://pub.dev/packages/langchain_ollama)       |           | ✔           | ✔         | [Ollama Chat API](https://ollama.ai) with tool-calling capabilities                                                                                                                                                                                                                                                                                                                                                                          |
 | [ChatOpenAI](https://langchaindart.dev/#/modules/model_io/models/chat_models/integrations/openai)                       | [langchain_openai](https://pub.dev/packages/langchain_openai)       | ✔         | ✔           | ✔         | [OpenAI Chat API](https://platform.openai.com/docs/api-reference/chat) and OpenAI Chat API compatible services ([TogetherAI](https://www.together.ai/), [Anyscale](https://www.anyscale.com/), [OpenRouter](https://openrouter.ai), [One API](https://github.com/songquanpeng/one-api), [Groq](https://groq.com/), [Llamafile](https://llamafile.ai/), [GPT4All](https://gpt4all.io/), [FastChat](https://github.com/lm-sys/FastChat), etc.) | 
 | [ChatVertexAI](https://langchaindart.dev/#/modules/model_io/models/chat_models/integrations/gcp_vertex_ai)              | [langchain_google](https://pub.dev/packages/langchain_google)       |           |             |           | [GCP Vertex AI Chat API](https://cloud.google.com/vertex-ai)                                                                                                                                                                                                                                                                                                                                                                                 |
