+++ packages/langchain_ollama/README.md
@@ -14,7 +14,6 @@
   * `Ollama`: wrapper around Ollama Completions API.
 - Chat models:
   * `ChatOllama`: wrapper around Ollama Chat API in a chat-like fashion.
-  * `ChatOllamaTools`: Wrapper around Ollama Chat API that enables tool calling capabilities.
 - Embeddings:
   * `OllamaEmbeddings`: wrapper around Ollama Embeddings API.
