+++ packages/langchain_ollama/lib/src/chat_models/chat_models.dart
@@ -1,4 +1,2 @@
 export 'chat_ollama/chat_ollama.dart';
 export 'chat_ollama/types.dart';
-export 'chat_ollama_tools/chat_ollama_tools.dart';
-export 'chat_ollama_tools/types.dart' hide conversationalResponseTool;
