+++ packages/langchain_ollama/test/chat_models/chat_ollama_test.dart
@@ -210,10 +210,10 @@ void main() {
 
     test('Test Multi-turn conversations', () async {
       final prompt = PromptValue.chat([
-        ChatMessage.humanText('List the numbers from 1 to 9 in order. '),
+        ChatMessage.humanText('List the numbers from 1 to 9 in order.'),
         ChatMessage.ai('123456789'),
         ChatMessage.humanText(
-          'Remove the number "4" from the list',
+          'Remove the number "4" from the list. Output only the remaining numbers in ascending order.',
         ),
       ]);
       final res = await chatModel.invoke(
@@ -253,31 +253,49 @@ void main() {
       expect(res.output.content.toLowerCase(), contains('apple'));
     });
 
+    const tool1 = ToolSpec(
+      name: 'get_current_weather',
+      description: 'Get the current weather in a given location',
+      inputJsonSchema: {
+        'type': 'object',
+        'properties': {
+          'location': {
+            'type': 'string',
+            'description': 'The city and country, e.g. San Francisco, US',
+          },
+          'unit': {
+            'type': 'string',
+            'enum': ['celsius', 'fahrenheit'],
+          },
+        },
+        'required': ['location'],
+      },
+    );
+    const tool2 = ToolSpec(
+      name: 'get_historic_weather',
+      description: 'Get the historic weather in a given location',
+      inputJsonSchema: {
+        'type': 'object',
+        'properties': {
+          'location': {
+            'type': 'string',
+            'description': 'The city and country, e.g. San Francisco, US',
+          },
+          'unit': {
+            'type': 'string',
+            'enum': ['celsius', 'fahrenheit'],
+          },
+        },
+        'required': ['location'],
+      },
+    );
+
     test('Test tool calling', timeout: const Timeout(Duration(minutes: 1)),
         () async {
-      const tool = ToolSpec(
-        name: 'get_current_weather',
-        description: 'Get the current weather in a given location',
-        inputJsonSchema: {
-          'type': 'object',
-          'properties': {
-            'location': {
-              'type': 'string',
-              'description': 'The city and country, e.g. San Francisco, US',
-            },
-            'unit': {
-              'type': 'string',
-              'description': 'The unit of temperature to return',
-              'enum': ['celsius', 'fahrenheit'],
-            },
-          },
-          'required': ['location'],
-        },
-      );
       final model = chatModel.bind(
         const ChatOllamaOptions(
           model: defaultModel,
-          tools: [tool],
+          tools: [tool1],
         ),
       );
 
@@ -290,13 +308,13 @@ void main() {
       expect(aiMessage1.toolCalls, hasLength(2));
 
       final toolCall1 = aiMessage1.toolCalls.first;
-      expect(toolCall1.name, tool.name);
+      expect(toolCall1.name, tool1.name);
       expect(toolCall1.arguments.containsKey('location'), isTrue);
       expect(toolCall1.arguments['location'], contains('Boston'));
       expect(toolCall1.arguments['unit'], 'celsius');
 
       final toolCall2 = aiMessage1.toolCalls.last;
-      expect(toolCall2.name, tool.name);
+      expect(toolCall2.name, tool1.name);
       expect(toolCall2.arguments.containsKey('location'), isTrue);
       expect(toolCall2.arguments['location'], contains('Madrid'));
       expect(toolCall2.arguments['unit'], 'celsius');
@@ -335,6 +353,59 @@ void main() {
       expect(aiMessage2.toolCalls, isEmpty);
       expect(aiMessage2.content, contains('22'));
       expect(aiMessage2.content, contains('25'));
+    });
+
+    test('Test multi tool call', () async {
+      final res = await chatModel.invoke(
+        PromptValue.string(
+          "What's the weather in Vellore, India and in Barcelona, Spain?",
+        ),
+        options: const ChatOllamaOptions(
+          model: defaultModel,
+          tools: [tool1, tool2],
+        ),
+      );
+      expect(res.output.toolCalls, hasLength(2));
+      final toolCall1 = res.output.toolCalls.first;
+      expect(toolCall1.name, 'get_current_weather');
+      expect(toolCall1.argumentsRaw, isNotEmpty);
+      expect(toolCall1.arguments, isNotEmpty);
+      expect(toolCall1.arguments['location'], 'Vellore, India');
+      expect(toolCall1.arguments['unit'], 'celsius');
+      final toolCall2 = res.output.toolCalls.last;
+      expect(toolCall2.name, 'get_current_weather');
+      expect(toolCall2.argumentsRaw, isNotEmpty);
+      expect(toolCall2.arguments, isNotEmpty);
+      expect(toolCall2.arguments['location'], 'Barcelona, Spain');
+      expect(toolCall2.arguments['unit'], 'celsius');
+      expect(res.finishReason, FinishReason.stop);
+    });
+
+    test('Test ChatToolChoice.none', () async {
+      final res = await chatModel.invoke(
+        PromptValue.string("What's the weather in Vellore, India?"),
+        options: const ChatOllamaOptions(
+          model: defaultModel,
+          tools: [tool1],
+          toolChoice: ChatToolChoice.none,
+        ),
+      );
+      expect(res.output.toolCalls, isEmpty);
+      expect(res.output.content, isNotEmpty);
+    });
+
+    test('Test ChatToolChoice.forced', () async {
+      final res = await chatModel.invoke(
+        PromptValue.string("What's the weather in Vellore, India?"),
+        options: ChatOllamaOptions(
+          model: defaultModel,
+          tools: const [tool1, tool2],
+          toolChoice: ChatToolChoice.forced(name: tool2.name),
+        ),
+      );
+      expect(res.output.toolCalls, hasLength(1));
+      final toolCall = res.output.toolCalls.first;
+      expect(toolCall.name, tool2.name);
     });
   });
 }
